requirements.txt
# Interfaz gráfica
PyQt6==6.6.1

# Base de datos MySQL
mysql-connector-python==9.0.0

# Reportes (Excel y PDF)
openpyxl==3.1.5
reportlab==4.2.2

# API REST
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2

# Seguridad
bcrypt==4.2.0
python-jose==3.3.0
passlib==1.7.4

# Configuración
python-dotenv==1.0.1

# Pruebas
pytest==8.3.3
pytest-qt==4.4.0

# Análisis de datos y gráficas (opcional)
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2

# Cliente HTTP
httpx==0.27.2

email-validator==2.0.0